App:
- Gerenciamento de usuários (administradores, secretárias, médicos).
- Login com UserName ou Email.

Testes:
- Testes do UserController... esses testes foram criados a partir de exceções reais que estavam ocorrendo.
      (ainda não fiz testes estensivos nesta parte do software... mas vou completar isso ainda)
- MockRepository: agora possui métodos para definir o usuário logado, a fim de permitir que sejam feitos testes
      com vários usuários, em várias condiçoes... inclusi criei um método para simular um usuário tentando acessar
      um practice que ele não deveria.

Firestarter:
- Separados alguns códigos de criação de dados em métodos.
- O CommandLine agora possui interação com usuário, e serve como uma ferramenta para manipular o banco.
- O CommandLine permite selecionar a conexão a ser usada, pois estava apontando hard-coded para o azure.

Outros:
- .gitignore: Permitir arquivos *.dll entre outros excluidos, só na pasta ThirdParty.
- Especificação alterada com alguns detalhes do funcionamento do login.
- Arquivos com as minhas micro-tarefas... é uma espécie de TODO meu (Miguel Angelo)
    tava mantendo isso só no meu computador, mas resolvi comitar... isso é legal.

+++ CerebelloWebRole/Areas/App/Models/PersonViewModels.cs
@@ -36,7 +36,7 @@ namespace CerebelloWebRole.Areas.App.Models
 
         [EnumDataType(typeof(TypeMaritalStatus))]
         [Display(Name = "Estado Civil")]
-        public int? MaritalStatus { get; set; }
+        public short? MaritalStatus { get; set; }
 
         [Display(Name = "Naturalidade")]
         public String BirthPlace { get; set; }
